Fix defects found reviewing the MCP server upgrade
Correctness:
- values at or after a nil were dropped from a tool result, so the Lua
  "return nil, reason" idiom reported success with no output at all
- numbers rendered at six significant digits, turning a room id or a
  timestamp into 1.23457e+06, and disagreeing with the same value nested
  in a table
- a numeric and a string key that render alike overwrote each other
- code was truncated at an embedded NUL, quietly running a shorter but
  still valid program
- luaTableToJson recursed without lua_checkstack, past LUA_MINSTACK

Security:
- readHost() also runs for an installed package, so a .mpackage could
  switch the MCP server on and open a Lua execution endpoint with no
  prompt. Only the profile's own file may enable it now.

Reporting:
- a bind clash cleared mEnableMCP, which XMLexport then wrote out, losing
  the setting for good. Every profile defaults to the same port, so this
  was the ordinary two-profile outcome. The reason now reaches the user
  and the preference survives.

Also drops a friend declaration on TLuaInterpreter that the public
getLuaGlobalState() already covers, an unused _meta constant and forward
declaration, and corrects comments that described code that is not there.

+++ src/Host.h
@@ -567,7 +567,7 @@ public:
     bool mEnableGMCP = true;
 #ifdef INCLUDE_MCPSERVER
     bool mEnableMCP = false;
-    int mMCPServerPort = 11235;
+    quint16 mMCPServerPort = 11235;
 #endif
     bool mEnableMSSP = true;
     bool mEnableMSDP = false;
